Redo credential link functionality

+++ app/src/main/java/com/harshnandwani/digitaltijori/presentation/credential/add_edit/AddEditCredentialActivity.kt
@@ -30,11 +30,11 @@ class AddEditCredentialActivity : ComponentActivity() {
             val mode = intent.getStringExtra(Parameters.KEY_MODE)
             if (mode == Parameters.VAL_MODE_ADD) {
                 val entity = intent.getSerializableExtra(Parameters.KEY_ENTITY) as Company?
-                val linkedAccountId = intent.getLongExtra(Parameters.KEY_BANK_ACCOUNT_ID, -1)
+                val linkedAccountId = intent.getIntExtra(Parameters.KEY_BANK_ACCOUNT_ID, -1)
                 if (entity != null) {
                     viewModel.onEvent(CredentialEvent.SelectEntity(entity))
                 }
-                if (linkedAccountId != -1L) {
+                if (linkedAccountId != -1) {
                     viewModel.onEvent(CredentialEvent.LinkToAccount(linkedAccountId))
                 }
             } else {
