+++ app/src/main/java/com/harshnandwani/digitaltijori/presentation/credential/add_edit/AddEditCredentialViewModel.kt
@@ -54,7 +54,7 @@ class AddEditCredentialViewModel @Inject constructor(
             is CredentialEvent.LinkToAccount -> {
                 _state.value.credential.value = state.value.credential.value.copy(
                     isLinkedToBank = true,
-                    bankAccountId = event.linkedAccountId.toInt()
+                    bankAccountId = event.linkedAccountId
                 )
             }
             is CredentialEvent.EnteredUsername -> {
